tools/legacy/logger-py/pyproject.toml
[project]
name = "logger"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ruff>=0.14.9",
    "pyserial>=3.5"
]
